refactor: rename page routers
