+++ web_ui/src/app/home/bots/components/bot-card/BotCardVO.ts
