+++ web_ui/src/app/home/bots/components/bot-form/ChooseAdapterEntity.ts
