+++ web_ui/src/app/home/models/component/llm-card/LLMCardVO.ts
