+++ web_ui/src/app/home/pipelines/components/pipeline-form/PipelineChildFormEntity.ts
